pyproject.toml
[tool.poetry]
name = "autosales"
version = "0.1.0"
description = ""
authors = ["Maksim Kozhin <a.zurnaaaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
django-rest-passwordreset = "^1.5.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
